refactor: colocate bundled memory providers under hermes_memory/plugins/memory
Move plugins/memory tree into hermes_memory.plugins.memory; update dynamic
loader to register hermes_memory.plugins.* parent packages and use
hermes_memory.plugins.memory.<name> module names for bundled providers.

Rewire run_agent, CLI, tests, and honcho imports to hermes_memory.plugins.memory.
Add plugins/memory/__init__.py shim re-exporting discovery API for backward
compatibility. Extend setuptools package find to hermes_memory.*.
