+++ hermes_memory/__init__.py
@@ -1,7 +1,8 @@
 """Hermes curated memory subsystem (file-backed MEMORY.md/USER.md + MemoryManager).
 
-External provider plugins remain under ``plugins/memory/``. This package holds
-the core implementation used by ``run_agent`` and the ``memory`` tool."""
+Bundled optional backends ship under ``hermes_memory/plugins/memory/``.
+User-installed providers still go in ``HERMES_HOME/plugins/<name>/``.
+This package holds the core implementation used by ``run_agent`` and the ``memory`` tool."""
 
 from hermes_memory.builtin_memory_tool import (
     MEMORY_SCHEMA,
