+++ hermes_memory/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Nested packages for optional Hermes integrations (e.g. memory providers)."""
