+++ hermes_memory/plugins/memory/byterover/plugin.yaml
