+++ hermes_memory/plugins/memory/hindsight/plugin.yaml
