+++ hermes_memory/plugins/memory/holographic/README.md
