+++ hermes_memory/plugins/memory/holographic/holographic.py
