+++ hermes_memory/plugins/memory/holographic/plugin.yaml
