+++ hermes_memory/plugins/memory/honcho/plugin.yaml
