+++ hermes_memory/plugins/memory/mem0/plugin.yaml
