+++ hermes_memory/plugins/memory/openviking/plugin.yaml
