+++ hermes_memory/plugins/memory/retaindb/plugin.yaml
