+++ hermes_memory/plugins/memory/supermemory/plugin.yaml
