+++ tests/honcho_plugin/test_async_memory.py
@@ -18,8 +18,8 @@ from unittest.mock import MagicMock, patch, call
 
 import pytest
 
-from plugins.memory.honcho.client import HonchoClientConfig
-from plugins.memory.honcho.session import (
+from hermes_memory.plugins.memory.honcho.client import HonchoClientConfig
+from hermes_memory.plugins.memory.honcho.session import (
     HonchoSession,
     HonchoSessionManager,
     _ASYNC_SHUTDOWN,
