+++ tests/plugins/memory/test_mem0_v2.py
@@ -6,7 +6,7 @@ Salvaged from PRs #5301 (qaqcvc) and #5117 (vvvanguards).
 import json
 import pytest
 
-from plugins.memory.mem0 import Mem0MemoryProvider
+from hermes_memory.plugins.memory.mem0 import Mem0MemoryProvider
 
 
 class FakeClientV2:
