+++ tests/plugins/memory/test_openviking_provider.py
@@ -1,7 +1,7 @@
 import json
 from unittest.mock import MagicMock
 
-from plugins.memory.openviking import OpenVikingMemoryProvider
+from hermes_memory.plugins.memory.openviking import OpenVikingMemoryProvider
 
 
 def test_tool_search_sorts_by_raw_score_across_buckets():
