+++ tests/test_honcho_client_config.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-from plugins.memory.honcho.client import HonchoClientConfig
+from hermes_memory.plugins.memory.honcho.client import HonchoClientConfig
 
 
 class TestHonchoClientConfigAutoEnable:
